refactor: reorganize SettingScene into settings package
