+++ core/src/main/java/com/p1_7/game/scenes/settings/VolumeSlider.java
@@ -1,0 +1,114 @@
+package com.p1_7.game.scenes.settings;
+
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Input;
+import com.badlogic.gdx.graphics.Color;
+import com.p1_7.abstractengine.entity.Entity;
+import com.p1_7.abstractengine.render.IDrawContext;
+import com.p1_7.abstractengine.render.IRenderable;
+import com.p1_7.abstractengine.transform.ITransform;
+import com.p1_7.game.core.Transform2D;
+import com.p1_7.game.input.ICursorSource;
+import com.p1_7.game.platform.GdxDrawContext;
+
+/**
+ * Horizontal volume slider entity.
+ */
+final class VolumeSlider extends Entity implements IRenderable {
+
+    private static final float KNOB_RADIUS = 14f;
+    private static final float TRACK_HALF_H = 5f;
+
+    private static final Color COLOUR_TRACK = new Color(0.20f, 0.20f, 0.55f, 1f);
+    private static final Color COLOUR_FILLED = new Color(0.80f, 0.80f, 1.00f, 1f);
+    private static final Color COLOUR_KNOB = new Color(0.35f, 0.35f, 0.80f, 1f);
+    private static final Color COLOUR_KNOB_DRAG = Color.WHITE;
+
+    private final Transform2D transform;
+    private final float trackLeft;
+    private final float trackCentreY;
+    private final float trackWidth;
+    private final float knobRadius;
+
+    private float value;
+    private boolean dragging = false;
+    private boolean moved = false;
+
+    VolumeSlider(float centreX, float centreY, float trackWidth, float initialValue) {
+        this.trackLeft = centreX - trackWidth / 2f;
+        this.trackCentreY = centreY;
+        this.trackWidth = trackWidth;
+        this.knobRadius = KNOB_RADIUS;
+        this.value = initialValue;
+        this.transform = new Transform2D(
+            trackLeft,
+            centreY - knobRadius,
+            trackWidth,
+            knobRadius * 2f
+        );
+    }
+
+    void updateInput(ICursorSource cursor) {
+        float mx = cursor.getCursorX();
+        float my = cursor.getCursorY();
+        float knobX = trackLeft + value * trackWidth;
+
+        moved = false;
+
+        if (Gdx.input.isButtonJustPressed(Input.Buttons.LEFT)) {
+            float dx = mx - knobX;
+            float dy = my - trackCentreY;
+            float hitRadius = knobRadius * 1.5f;
+            if (dx * dx + dy * dy <= hitRadius * hitRadius) {
+                dragging = true;
+            }
+        }
+
+        if (dragging && Gdx.input.isButtonPressed(Input.Buttons.LEFT)) {
+            float clamped = Math.max(trackLeft, Math.min(trackLeft + trackWidth, mx));
+            value = (clamped - trackLeft) / trackWidth;
+            moved = true;
+        }
+
+        if (!Gdx.input.isButtonPressed(Input.Buttons.LEFT)) {
+            dragging = false;
+        }
+    }
+
+    float getValue() {
+        return value;
+    }
+
+    boolean hasMoved() {
+        return moved;
+    }
+
+    void resetMoved() {
+        moved = false;
+    }
+
+    @Override
+    public String getAssetPath() {
+        return null;
+    }
+
+    @Override
+    public ITransform getTransform() {
+        return transform;
+    }
+
+    @Override
+    public void render(IDrawContext ctx) {
+        GdxDrawContext gdxCtx = (GdxDrawContext) ctx;
+        float knobX = trackLeft + value * trackWidth;
+
+        gdxCtx.rect(COLOUR_TRACK, trackLeft, trackCentreY - TRACK_HALF_H,
+            trackWidth, TRACK_HALF_H * 2f, true);
+        gdxCtx.rect(COLOUR_FILLED, trackLeft, trackCentreY - TRACK_HALF_H,
+            value * trackWidth, TRACK_HALF_H * 2f, true);
+        gdxCtx.circle(dragging ? COLOUR_KNOB_DRAG : COLOUR_KNOB,
+            knobX, trackCentreY, knobRadius, true);
+    }
+
+    void dispose() { }
+}
